apis/mongo/python/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
arrow==1.2.2
    # via -r .\requirements.in
azure-core==1.11.0
    # via azure-cosmos
azure-cosmos==4.2.0
    # via -r .\requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
docopt==0.6.2
    # via -r .\requirements.in
flake8==3.8.4
    # via -r .\requirements.in
idna==2.10
    # via requests
mccabe==0.6.1
    # via flake8
pipdeptree==2.2.1
    # via -r .\requirements.in
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pymongo==3.11.3
    # via -r .\requirements.in
python-dateutil==2.8.2
    # via arrow
requests==2.25.1
    # via azure-core
six==1.15.0
    # via
    #   azure-core
    #   azure-cosmos
    #   python-dateutil
urllib3==1.26.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
